% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{discly}
\alias{discly}
\title{discounted life-years}
\usage{
discly(
  iso3,
  age,
  yearnow,
  sex = "Total",
  endyear = 2098,
  HR = 1,
  dr = 0,
  hiv = "both"
)
}
\arguments{
\item{iso3}{country}

\item{age}{age now}

\item{yearnow}{year now}

\item{sex}{Female/Male/Total}

\item{endyear}{end year}

\item{HR}{hazard ratio of death if relevant}

\item{dr}{discount rate (exponential version)}

\item{hiv}{HIV specific estimates? ('both'/'hiv+'/'hiv-')}
}
\value{
discounted life years
}
\description{
Discounted life years
}
\details{
This calculates discounted life-years, potentially allowing a mortality hazard ratio. It also allows approximate HIV-specific estimates. See README for reference to methodology.
}
\examples{
discly('ZWE',5,2020)
discly('ZWE',5,2020,dr=0.03)
}
\author{
Pete Dodd
}
