% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCNdata}
\alias{getCNdata}
\title{Get input data}
\usage{
getCNdata(cn, sx)
}
\arguments{
\item{cn}{iso3 code for country}

\item{sx}{sex = Female/Male/Total}
}
\value{
matrix ages, years
}
\description{
Get matrix from data in right format
}
\author{
Pete Dodd
}
