% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCNdataH}
\alias{getCNdataH}
\title{Get input data}
\usage{
getCNdataH(cn, sx, h = "hiv+")
}
\arguments{
\item{cn}{iso3 code for country}

\item{sx}{sex = Female/Male/Total}
}
\value{
matrix ages, years
}
\description{
Get matrix from data in right format (HIV stratified)
}
\author{
Pete Dodd
}
