% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getHzhist}
\alias{getHzhist}
\title{Projected death risks}
\usage{
getHzhist(yr0, age0, yr1, M)
}
\arguments{
\item{yr0}{initial year}

\item{age0}{initial age}

\item{yr1}{final year}

\item{M}{matrix to extrapolate}
}
\value{
vector hazard for years
}
\description{
Get diagonal hazard
}
\details{
TODO
}
\author{
Pete Dodd
}
